requirements.txt
ping3
python-dotenv
requests
